Change code structure to one `app.py`
